// typechain-types/SimpleAgentRegistry.ts
export interface SimpleAgentRegistryInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "getAgent"
      | "getReputation"
      | "getSignal"
      | "registerAgent"
      | "resolveAgent"
      | "submitSignal"
  ): FunctionFragment;

  getEvent(
    nameOrSignatureOrTopic: "AgentRegistered" | "SignalSubmitted"
  ): EventFragment;

  encodeFunctionData(
    functionFragment: "getAgent",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "getReputation",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "getSignal",
    values: [BigNumberish, BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "registerAgent",
    values: [AddressLike, string, string]
  ): string;
  encodeFunctionData(
    functionFragment: "resolveAgent",
    values: [string]
  ): string;
  encodeFunctionData(
    functionFragment: "submitSignal",
    values: [BigNumberish, string, BigNumberish, string]
  ): string;

  decodeFunctionResult(functionFragment: "getAgent", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "getReputation",
    data: BytesLike
  ): Result;
  decodeFunctionResult(functionFragment: "getSignal", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "registerAgent",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "resolveAgent",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "submitSignal",
    data: BytesLike
  ): Result;
}

export namespace AgentRegisteredEvent {
  export type InputTuple = [
    agentId: BigNumberish,
    agent: AddressLike,
    did: string
  ];
  export type OutputTuple = [agentId: bigint, agent: string, did: string];
  export interface OutputObject {
    agentId: bigint;
    agent: string;
    did: string;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace SignalSubmittedEvent {
  export type InputTuple = [
    agentId: BigNumberish,
    signalType: string,
    value: BigNumberish
  ];
  export type OutputTuple = [
    agentId: bigint,
    signalType: string,
    value: bigint
  ];
  export interface OutputObject {
    agentId: bigint;
    signalType: string;
    value: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface SimpleAgentRegistry extends BaseContract {
  connect(runner?: ContractRunner | null): SimpleAgentRegistry;
  waitForDeployment(): Promise<this>;

  interface: SimpleAgentRegistryInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  getAgent: TypedContractMethod<
    [agentId: BigNumberish],
    [
      [string, string, string] & {
        agent: string;
        did: string;
        metadata: string;
      }
    ],
    "view"
  >;

  getReputation: TypedContractMethod<
    [agentId: BigNumberish],
    [[bigint, bigint] & { score: bigint; signalCount: bigint }],
    "view"
  >;

  getSignal: TypedContractMethod<
    [agentId: BigNumberish, index: BigNumberish],
    [
      [string, bigint, string, bigint] & {
        signalType: string;
        value: bigint;
        metadata: string;
        timestamp: bigint;
      }
    ],
    "view"
  >;

  registerAgent: TypedContractMethod<
    [agent: AddressLike, did: string, metadata: string],
    [bigint],
    "nonpayable"
  >;

  resolveAgent: TypedContractMethod<[did: string], [bigint], "view">;

  submitSignal: TypedContractMethod<
    [
      agentId: BigNumberish,
      signalType: string,
      value: BigNumberish,
      metadata: string
    ],
    [void],
    "nonpayable"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "getAgent"
  ): TypedContractMethod<
    [agentId: BigNumberish],
    [
      [string, string, string] & {
        agent: string;
        did: string;
        metadata: string;
      }
    ],
    "view"
  >;
  getFunction(
    nameOrSignature: "getReputation"
  ): TypedContractMethod<
    [agentId: BigNumberish],
    [[bigint, bigint] & { score: bigint; signalCount: bigint }],
    "view"
  >;
  getFunction(
    nameOrSignature: "getSignal"
  ): TypedContractMethod<
    [agentId: BigNumberish, index: BigNumberish],
    [
      [string, bigint, string, bigint] & {
        signalType: string;
        value: bigint;
        metadata: string;
        timestamp: bigint;
      }
    ],
    "view"
  >;
  getFunction(
    nameOrSignature: "registerAgent"
  ): TypedContractMethod<
    [agent: AddressLike, did: string, metadata: string],
    [bigint],
    "nonpayable"
  >;
  getFunction(
    nameOrSignature: "resolveAgent"
  ): TypedContractMethod<[did: string], [bigint], "view">;
  getFunction(
    nameOrSignature: "submitSignal"
  ): TypedContractMethod<
    [
      agentId: BigNumberish,
      signalType: string,
      value: BigNumberish,
      metadata: string
    ],
    [void],
    "nonpayable"
  >;

  getEvent(
    key: "AgentRegistered"
  ): TypedContractEvent<
    AgentRegisteredEvent.InputTuple,
    AgentRegisteredEvent.OutputTuple,
    AgentRegisteredEvent.OutputObject
  >;
  getEvent(
    key: "SignalSubmitted"
  ): TypedContractEvent<
    SignalSubmittedEvent.InputTuple,
    SignalSubmittedEvent.OutputTuple,
    SignalSubmittedEvent.OutputObject
  >;

  filters: {
    "AgentRegistered(uint256,address,string)": TypedContractEvent<
      AgentRegisteredEvent.InputTuple,
      AgentRegisteredEvent.OutputTuple,
      AgentRegisteredEvent.OutputObject
    >;
    AgentRegistered: TypedContractEvent<
      AgentRegisteredEvent.InputTuple,
      AgentRegisteredEvent.OutputTuple,
      AgentRegisteredEvent.OutputObject
    >;

    "SignalSubmitted(uint256,string,int256)": TypedContractEvent<
      SignalSubmittedEvent.InputTuple,
      SignalSubmittedEvent.OutputTuple,
      SignalSubmittedEvent.OutputObject
    >;
    SignalSubmitted: TypedContractEvent<
      SignalSubmittedEvent.InputTuple,
      SignalSubmittedEvent.OutputTuple,
      SignalSubmittedEvent.OutputObject
    >;
  };
}